refactor: [M3-6417] - SRC > Components > Tile
