+++ packages/manager/src/components/Tile/Tile.stories.mdx
@@ -1,47 +1,46 @@
 import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import Tile from 'src/components/Tile';
+import { Tile } from 'src/components/Tile/Tile';
 import Chat from 'src/assets/icons/chat.svg';
 
-<Meta title="Components/Tile" component={Tile}/>
+<Meta title="Components/Tile" component={Tile} />
 
 export const TileTemplate = (args, context) => <Tile {...args} />;
-
-
 
 # Tile
 
 <Canvas>
-    <Story
-        name="Tile"
-        args={{
-            title: 'This is the Tile title',
-            description: "In order to make the tile a link, the link prop needs to be set. It can be either an internal or external link, or an onClick function",
-            link: "http://cloud.linode.com",
-            icon: (<Chat />),
-            errorText: undefined,
-        }}
-        argTypes={{
-            title: {
-                description: 'A string to display in the title section of the Tile.',
-            },
-            description: {
-                description: 'A string to display in the body of the Tile.',
-            },
-            link: {
-                description: 'A url to send the user to if they click on the Tile.'
-            },
-            errorText: {
-                type: 'string',
-                description: 'A string to display in an error notice above the title.',
-            },
-            icon: {
-                description: 'An SVG to display at the top of the Tile.',
-                control: false,
-            },
-        }}
-    >
-        {TileTemplate.bind({})}
-    </Story>
+  <Story
+    name="Tile"
+    args={{
+      title: 'This is the Tile title',
+      description:
+        'In order to make the tile a link, the link prop needs to be set. It can be either an internal or external link, or an onClick function',
+      link: 'http://cloud.linode.com',
+      icon: <Chat />,
+      errorText: undefined,
+    }}
+    argTypes={{
+      title: {
+        description: 'A string to display in the title section of the Tile.',
+      },
+      description: {
+        description: 'A string to display in the body of the Tile.',
+      },
+      link: {
+        description: 'A url to send the user to if they click on the Tile.',
+      },
+      errorText: {
+        type: 'string',
+        description: 'A string to display in an error notice above the title.',
+      },
+      icon: {
+        description: 'An SVG to display at the top of the Tile.',
+        control: false,
+      },
+    }}
+  >
+    {TileTemplate.bind({})}
+  </Story>
 </Canvas>
 
 <ArgsTable story="Tile" sort="requiredFirst" exclude={['className']} />
