--- packages/manager/src/components/Tile/index.ts
@@ -1,2 +1,0 @@
-import Tile from './Tile';
-export default Tile;
